requirements.txt
python-telegram-bot==20.8
openai==1.93.0
aiogram==3.5.0
python-dotenv==1.0.0
requests==2.31.0
sqlalchemy==2.0.25
python-multipart==0.0.6
uvicorn==0.27.0
fastapi==0.109.0
pydantic==2.5.3
